+++ example/spec/helpers/granuleUtils.js
@@ -214,7 +214,7 @@ const waitForGranuleRecordUpdatedInList = async (stackName, granule, additionalQ
       'beginningDateTime',
       'endingDateTime',
       'error',
-      'files', // FUTURE: files contain a "type" field in elasticsearch that does not exist in Postgres
+      'files', // TODO -2714 this should be removed
       'lastUpdateDateTime',
       'productionDateTime',
       'updatedAt',
@@ -230,6 +230,7 @@ const waitForGranuleRecordUpdatedInList = async (stackName, granule, additionalQ
     });
     const results = JSON.parse(resp.body).results;
     if (results && results.length === 1) {
+      // TODO - CUMULUS-2714 key sort both files objects for comparison
       const granuleMatches = isEqual(omit(results[0], fieldsIgnored), omit(granule, fieldsIgnored));
 
       if (!granuleMatches) {