+++ packages/db/src/translate/file.ts
@@ -27,6 +27,7 @@ export const translatePostgresFileToApiFile = (
   key: filePgRecord.key,
   size: filePgRecord.file_size ? Number.parseInt(filePgRecord.file_size, 10) : undefined,
   source: filePgRecord.source,
+  type: filePgRecord.type,
 });
 
 export const translateApiFiletoPostgresFile = (
@@ -44,5 +45,6 @@ export const translateApiFiletoPostgresFile = (
     file_size: file.size,
     path: file.path,
     source: file.source,
+    type: file.type,
   };
 };